+++ moon/Status.cpp
@@ -29,8 +29,8 @@ const char* Status::copyState(const char *state) {
 }
 
 Status::Status(int code, const Slice& msg, const Slice& msg2) {
-    const uint32_t len1 = msg.size();
-    const uint32_t len2 = msg2.size();
+    const uint32_t len1 = static_cast<uint32_t>(msg.size());
+    const uint32_t len2 = static_cast<uint32_t>(msg2.size());
     const uint32_t size = len1 + (len2 ? (2 + len2) : 0);
     char *result = new char[size + 6];
     memcpy(result, &size, sizeof(size));
